move StoreInvoiceStub to it's own file
