+++ test/invoice/create_invoice_test.dart
@@ -5,6 +5,7 @@ import 'package:thepos/features/invoice/data/models/invoice_item.dart';
 import 'package:thepos/features/invoice/data/repositories/store_invoice_repository.dart';
 
 import 'helpers/shared_test_helper.dart';
+import 'helpers/store_invoice_stub.dart';
 
 void main() {
   StoreInvoiceRepository makeSUT(
@@ -54,15 +55,4 @@ void main() {
 
     expectInvoice(result, remoteResult);
   });
-}
-
-class StoreInvoiceStub extends StoreInvoice {
-  StoreInvoiceStub(this._result);
-
-  final Invoice _result;
-
-  @override
-  Future<Invoice> store(Invoice invoice) {
-    return Future<Invoice>.value(_result);
-  }
 }
